check in code

+++ FangZhouShuMa.ApplicationCore/Interfaces/Services/IBasketService.cs
@@ -7,7 +7,7 @@ namespace FangZhouShuMa.ApplicationCore.Interfaces
     {
         Task<decimal> GetBasketItemCountAsync(string userName);
         Task TransferBasketAsync(string anonymousId, string userName);
-        Task AddItemToBasket(int basketId, int productId, decimal price, int quantity);
+        Task AddItemToBasket(int basketId, int productId, decimal price, decimal quantity);
         Task SetQuantities(int basketId, Dictionary<string, int> quantities);
         Task DeleteBasketAsync(int basketId);
     }
